requirements.txt
flet~=0.10.1
docker~=6.1.3
PyYAML~=6.0.1
requests~=2.31.0
setuptools~=65.5.1
Jinja2~=3.1.2